Move new events routes to calendar routes
