+++ src/Module/Calendar/Export.php
@@ -19,7 +19,7 @@
  *
  */
 
-namespace Friendica\Module\Events;
+namespace Friendica\Module\Calendar;
 
 use Friendica\App;
 use Friendica\BaseModule;
@@ -34,7 +34,7 @@ use Friendica\Util\Profiler;
 use Psr\Log\LoggerInterface;
 
 /**
- * Controller to export calendar
+ * Controller to export a calendar from a given user
  */
 class Export extends BaseModule
 {
@@ -88,7 +88,7 @@ class Export extends BaseModule
 
 		// If nothing went wrong we can echo the export content
 		if ($evexport["success"]) {
-			$this->response->setHeader(sprintf('content-disposition: attachment; filename="%s-%s.%s"',
+			$this->response->setHeader(sprintf('Content-Disposition: attachment; filename="%s-%s.%s"',
 				$this->t('calendar'),
 				$this->parameters['nickname'],
 				$evexport["extension"]
