+++ static/routes.config.php
@@ -365,6 +365,9 @@ return [
 
 	'/bookmarklet'         => [Module\Bookmarklet::class,  [R::GET]],
 
+	'/calendar/{nickname}/export[/{format}]' => [Module\Calendar\Export::class, [R::GET]],
+	'/calendar/json'                           => [Module\Calendar\Json::class, [R::GET]],
+
 	'/community[/{content}]' => [Module\Conversation\Community::class, [R::GET]],
 
 	'/compose[/{type}]'    => [Module\Item\Compose::class, [R::GET, R::POST]],
@@ -397,9 +400,6 @@ return [
 	'/dfrn_poll/{nickname}'     => [Module\DFRN\Poll::class,        [R::GET]],
 	'/dirfind'                  => [Module\Search\Directory::class, [R::GET]],
 	'/directory'                => [Module\Directory::class,        [R::GET]],
-
-	'/events/{nickname}/export[/{format}]' => [Module\Events\Export::class, [R::GET]],
-	'/events/json'                         => [Module\Events\Json::class,   [R::GET]],
 
 	'/featured/{nickname}'      => [Module\ActivityPub\Featured::class, [R::GET]],
