+++ view/templates/widget/events.tpl
@@ -3,7 +3,7 @@
 	<h3>{{$etitle}}</h3>
 
 	<ul class="sidebar-calendar-export-ul">
-		<li role="menuitem" class="sidebar-calendar-export-li"><a href="{{$baseurl}}/events/{{$user}}/export/ical">{{$export_ical}}</a></li>
-		<li role="menuitem" class="sidebar-calendar-export-li"><a href="{{$baseurl}}/events/{{$user}}/export/csv">{{$export_csv}}</a></li>
+		<li role="menuitem" class="sidebar-calendar-export-li"><a href="{{$baseurl}}/calendar/{{$user}}/export/ical">{{$export_ical}}</a></li>
+		<li role="menuitem" class="sidebar-calendar-export-li"><a href="{{$baseurl}}/calendar/{{$user}}/export/csv">{{$export_csv}}</a></li>
 	</ul>
 </div>
